Use VTS fork of Wisper We've forked wisper to include two concepts that are not included in the latest release. The first is to remove remove the bin/executable binaries. The bin/executable binaries clash with the latest version of fog-aws.
The second is to make wisper compabtible with ruby 3. We've cherry picked from an open PR that make the gem ruby 3 compatible. Here is a link to that branch: https://github.com/doits/wisper/tree/ruby_3_compat The hope is that this will set us up to upgrade VTS core to ruby 3.
[INFORM-2458]