Remove dependency on wisper
[INFORM-3178](https://viewthespace.atlassian.net/browse/INFORM-3178)

wisper prevents upgrading to Ruby 3.0. We cannot place it in a Gemfile
with our specified version of wisper because it will not be a dependency
of a gem and therefore will not be installed when tty-reader is
installed.

Instead, we place this version in a Gemfile along with tty-reader so it
is available to it.